+++ safe_core/src/ffi/error_codes.rs
@@ -74,6 +74,7 @@ pub const ERR_INVALID_PERMISSIONS_SUCCESSOR: i32 = -3002;
 pub const ERR_SIGN_KEYTYPE_MISMATCH: i32 = -3003;
 pub const ERR_INVALID_SIGNATURE: i32 = -3004;
 pub const ERR_INVALID_CAVEATS: i32 = -3005;
+pub const ERR_INVALID_TOKEN: i32 = -3006;
 
 // Coin errors.
 pub const ERR_LOSS_OF_PRECISION: i32 = -4000;