+++ safe_core/src/lib.rs
@@ -39,9 +39,9 @@ pub use ffi::*;
 
 // Export public core interface.
 
-pub use self::client::{
-    mdata_info, recoverable_apis, test_create_balance, AuthActions, Client, ClientKeys, MDataInfo,
-};
+#[cfg(any(test, feature = "testing"))]
+pub use self::client::test_create_balance;
+pub use self::client::{mdata_info, recoverable_apis, AuthActions, Client, ClientKeys, MDataInfo};
 #[cfg(feature = "mock-network")]
 pub use self::client::{mock_vault_path, MockConnectionManager as ConnectionManager};
 pub use self::config_handler::config_dir;